Combine serializer and deserializer
