+++ sdks/java/core/src/main/java/org/apache/beam/sdk/util/RowJsonUtils.java
@@ -24,14 +24,17 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import java.io.IOException;
 import org.apache.beam.sdk.annotations.Internal;
-import org.apache.beam.sdk.util.RowJsonDeserializer.UnsupportedRowJsonException;
+import org.apache.beam.sdk.util.RowJson.RowJsonDeserializer.UnsupportedRowJsonException;
 import org.apache.beam.sdk.values.Row;
 
-/** Utilities for working with {@link RowJsonSerializer} and {@link RowJsonDeserializer}. */
+/**
+ * Utilities for working with {@link RowJson.RowJsonSerializer} and {@link
+ * RowJson.RowJsonDeserializer}.
+ */
 @Internal
 public class RowJsonUtils {
 
-  public static ObjectMapper newObjectMapperWith(RowJsonDeserializer deserializer) {
+  public static ObjectMapper newObjectMapperWith(RowJson.RowJsonDeserializer deserializer) {
     SimpleModule module = new SimpleModule("rowDeserializationModule");
     module.addDeserializer(Row.class, deserializer);
 
@@ -41,7 +44,7 @@ public class RowJsonUtils {
     return objectMapper;
   }
 
-  public static ObjectMapper newObjectMapperWith(RowJsonSerializer serializer) {
+  public static ObjectMapper newObjectMapperWith(RowJson.RowJsonSerializer serializer) {
     SimpleModule module = new SimpleModule("rowSerializationModule");
     module.addSerializer(Row.class, serializer);
