+++ sdks/java/core/src/main/java/org/apache/beam/sdk/util/RowJsonValidation.java
@@ -60,7 +60,7 @@ class RowJsonValidation {
     }
 
     if (!SUPPORTED_TYPES.contains(fieldTypeName)) {
-      throw new RowJsonDeserializer.UnsupportedRowJsonException(
+      throw new RowJson.RowJsonDeserializer.UnsupportedRowJsonException(
           fieldTypeName.name()
               + " is not supported when converting JSON objects to Rows. "
               + "Supported types are: "
